requirements.txt
python-telegram-bot[job-queue]==21.6
APScheduler==3.10.4
python-dotenv==1.0.1
telethon==1.36.0
